AIE5-DeployPythonicRAG/fastapi-rag2/requirements.txt
fastapi
python-multipart
uvicorn
python-dotenv
openai
pydantic
pypdf
PyPDF2
langchain
faiss-cpu 